upate z80core for memory config from system.conf
