+++ z80core/simglb.c
@@ -145,6 +145,8 @@ int f_flag;			/* flag for -f option */
 int u_flag;			/* flag for -u option */
 int r_flag;			/* flag for -r option */
 int c_flag;			/* flag for -c option */
+int M_flag = 0;		/* flag for -M option */
+int R_flag = 0;		/* flag for -R option */
 
 /*
  *	Variables for configuration and disk images
